feat: Add thumbnail support for file uploads
The code changes introduce a new `thumbnailBuffer` field in the `fileSchema` of the `upload.js` model. This field will store the thumbnail image buffer if the uploaded file is an allowed image type. The `fileType` field is also added to store the MIME type of the file. Additionally, the code modifies the `/upload` route in `file.js` to generate and save the thumbnail buffer when processing an allowed image type file. This enhancement improves the user experience by providing thumbnail previews for image files during file uploads.